+++ src/main/java/roomescape/client/feign/PaymentFeignConfig.java
@@ -1,4 +1,4 @@
-package roomescape.client;
+package roomescape.client.feign;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
